+++ src/common/js/index.js
@@ -4,8 +4,8 @@
 
 import './global.js';
 import './constants.js';
+import './verbose.js';
 //
 import './configure.js';
 import './ready.js';
 import './startup.js';
-import './verbose.js';