+++ ChangeLog.md
@@ -7,6 +7,8 @@
     Release date: 
 
     - The package is extended to the client side to let public settings land to the client (thus bumping the candidate release number)
+    - No more wait for Meteor.startup() to load private configuration assets (todo #7)
+      the verbosity configuration becomes obsolete, this can be controlled through the hardcoded EnvSettings.C.WaitForStartup constant
 
 ### 1.6.1